fix bert
